Spring Boot 4.0.1, Java 21, and Spring Security 7.x  Upgraded. Running application

+++ src/main/java/com/javatechie/crud/example/entity/Product.java
@@ -18,8 +18,8 @@ public class Product {
 
     @Id
     @GeneratedValue
-    private int id;
+    private Integer  id;
     private String name;
-    private int quantity;
-    private double price;
+    private Integer  quantity;
+    private Integer  price;
 }
